+++ database/factories/BillFactory.php
@@ -1,0 +1,28 @@
+<?php
+
+namespace Database\Factories;
+
+use App\Models\Bill;
+use Illuminate\Database\Eloquent\Factories\Factory;
+
+class BillFactory extends Factory
+{
+    /**
+     * The name of the factory's corresponding model.
+     *
+     * @var string
+     */
+    protected $model = Bill::class;
+
+    /**
+     * Define the model's default state.
+     *
+     * @return array
+     */
+    public function definition()
+    {
+        return [
+            'bill_name' => $this->faker->text,
+        ];
+    }
+}
