+++ database/factories/ComponentFactory.php
@@ -1,0 +1,29 @@
+<?php
+
+namespace Database\Factories;
+
+use App\Models\Component;
+use Illuminate\Database\Eloquent\Factories\Factory;
+
+class ComponentFactory extends Factory
+{
+    /**
+     * The name of the factory's corresponding model.
+     *
+     * @var string
+     */
+    protected $model = Component::class;
+
+    /**
+     * Define the model's default state.
+     *
+     * @return array
+     */
+    public function definition()
+    {
+        return [
+            'device_id' => $this->faker->randomDigit,
+            'type' => $this->faker->regexify('[A-Za-z0-9]{4,20}'),
+        ];
+    }
+}
